module exposed via discovery function
