+++ .eslintignore
@@ -1,4 +1,2 @@
 templates/
-common-templates/
 node_modules/
-playground/
