+++ package.json
@@ -46,7 +46,7 @@
     "js-yaml": "^4.1.0"
   },
   "dependencies": {
-    "@adobe/generator-app-common-lib": "0.0.0",
+    "@adobe/generator-app-common-lib": "https://github.com/adobe/generator-app-common-lib.git#27123e57d2d6ca891bcb1f85ed8eec4996269acc",
     "upath": "^2.0.1",
     "yeoman-generator": "^5.4.2"
   }
