+++ templates/add-action/asset-compute/templates/.npmignore
