+++ templates/add-action/asset-compute/templates/_worker.js
