+++ templates/add-action/asset-compute/templates/test/corrupt-input/params.json
