+++ templates/add-action/asset-compute/templates/test/simple-test/rendition.jpg
